유저정보 수정에서 기존파일 null 체크 & 회원가입 비밀번호 유효성 체크

+++ src/main/java/freshtrash/freshtrashbackend/controller/MemberApi.java
@@ -6,6 +6,7 @@ import freshtrash.freshtrashbackend.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -35,14 +36,14 @@ public class MemberApi {
     public ResponseEntity<MemberResponse> updateMember(
             @AuthenticationPrincipal MemberPrincipal memberPrincipal,
             @RequestPart MemberRequest memberRequest,
-            @RequestPart MultipartFile imgFile) {
+            @RequestPart(required = false) MultipartFile imgFile) {
         String oldFile =
                 memberService.findFileNameOfMember(memberPrincipal.id()).fileName();
         MemberResponse memberResponse =
                 MemberResponse.fromEntity(memberService.updateMember(memberPrincipal.id(), memberRequest, imgFile));
 
         // 파일이 수정된 경우 -> 이전 파일 삭제
-        if (!oldFile.equals(memberResponse.fileName())) {
+        if (StringUtils.hasText(oldFile) && !oldFile.equals(memberResponse.fileName())) {
             memberService.deleteOldFile(oldFile);
         }
         return ResponseEntity.ok(memberResponse);
